+++ src/main/java/dev/peertosir/ownchaincrm/repository/DetailRepository.java
@@ -3,4 +3,4 @@ package dev.peertosir.ownchaincrm.repository;
 import dev.peertosir.ownchaincrm.domain.Detail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetailRepository extends JpaRepository<Detail, Integer> {}
+public interface DetailRepository extends JpaRepository<Detail, Long> {}
