+++ src/main/java/dev/peertosir/ownchaincrm/repository/SchemaRepository.java
@@ -3,5 +3,5 @@ package dev.peertosir.ownchaincrm.repository;
 import dev.peertosir.ownchaincrm.domain.Schema;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SchemaRepository extends JpaRepository<Schema, Integer> {
+public interface SchemaRepository extends JpaRepository<Schema, Long> {
 }
